(PC-33723)[API] chore: move RedactorInformation to core/educational/schemas.py
