+++ api/src/pcapi/core/educational/schemas.py
@@ -1,0 +1,9 @@
+from pcapi.routes.serialization import BaseModel
+
+
+class RedactorInformation(BaseModel):
+    civility: str | None
+    lastname: str | None
+    firstname: str | None
+    email: str
+    uai: str
